% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIMST-package.R
\docType{package}
\name{MSIMST}
\alias{MSIMST-package}
\alias{MSIMST}
\title{The 'MSIMST' package.}
\description{
Incorporates a Bayesian monotonic single-index mixed-effect model with a multivariate skew-t likelihood, specifically designed to handle survey weights adjustments. Features include a simulation program and an associated Gibbs sampler for model estimation. The single-index function is constrained to be monotonic increasing, utilizing a customized Gaussian process prior for precise estimation. The model assumes random effects follow a canonical skew-t distribution, while residuals are represented by a multivariate Student-t distribution. Offers robust Bayesian adjustments to integrate survey weight information effectively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rh8liuqy/MSIMST}
}

}
\author{
\strong{Maintainer}: Qingyang Liu \email{rh8liuqy@gmail.com} (\href{https://orcid.org/0000-0003-3265-6330}{ORCID})

Authors:
\itemize{
  \item Debdeep Pati \email{debdeep@stat.tamu.edu}
  \item Dipankar Bandyopadhyay \email{dbandyop@vcu.edu}
}

}
