% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIMST-package.R
\docType{package}
\name{MSIMST}
\alias{MSIMST-package}
\alias{MSIMST}
\title{The 'MSIMST' package.}
\description{
A Bayesian monotonic single-index mixed-effect model incorporating a multivariate skew-t likelihood with survey weights adjustments. This package includes a simulation program and the associated Gibbs sampler. The single-index function is modeled as a monotonic increasing function, with a tailored Gaussian process prior to ensure accurate estimation. Random effects are assumed to follow the canonical skew-t distribution, while residuals are modeled using the multivariate Student-t distribution. Additionally, the package provides Bayesian adjustment for survey weight information.
}
\author{
\strong{Maintainer}: Qingyang Liu \email{rh8liuqy@gmail.com} (\href{https://orcid.org/0000-0003-3265-6330}{ORCID})

Authors:
\itemize{
  \item Debdeep Pati \email{debdeep@stat.tamu.edu}
  \item Dipankar Bandyopadhyay \email{dbandyop@vcu.edu}
}

}
